[backend] Move code from framework to api
