+++ openbas-api/src/main/java/io/openbas/executors/caldera/client/model/Link.java
