+++ openbas-api/src/main/java/io/openbas/executors/caldera/client/model/Result.java
