+++ openbas-api/src/main/java/io/openbas/executors/tanium/TaniumExecutor.java
@@ -9,7 +9,6 @@ import io.openbas.integrations.ExecutorService;
 import io.openbas.integrations.InjectorService;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 import org.springframework.stereotype.Service;
