+++ openbas-api/src/main/java/io/openbas/executors/tanium/config/TaniumExecutorConfig.java
