+++ openbas-api/src/main/java/io/openbas/executors/tanium/model/Endpoints.java
