+++ openbas-api/src/main/java/io/openbas/utils/TemplateHelper.java
@@ -1,8 +1,9 @@
-package io.openbas.helper;
+package io.openbas.utils;
 
 import freemarker.template.Configuration;
 import freemarker.template.Template;
 import io.openbas.execution.ExecutionContext;
+import io.openbas.helper.TemplateExceptionManager;
 import org.springframework.ui.freemarker.FreeMarkerTemplateUtils;
 
 import java.io.StringReader;
