--- openbas-framework/src/main/java/io/openbas/executors/README.md
@@ -1,1 +1,0 @@
-# Built-in executors
